% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gbrs-package}
\alias{gbrs-package}
\title{gbrs: Gradient Boosted Risk Scores}
\description{
Fits interpretable risk scores via gradient boosting for regression, binary classification, and survival outcomes. Provides a formula interface, bootstrap confidence intervals, and serialization to a portable JSON format that can be re-loaded from Python.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cogeor/GBRS}
  \item Report bugs at \url{https://github.com/cogeor/GBRS/issues}
}

}
\author{
\strong{Maintainer}: Costa Georgantas \email{costa.georgantas@chuv.ch}

Authors:
\itemize{
  \item Costa Georgantas \email{costa.georgantas@chuv.ch}
}

}
\keyword{internal}
