% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_io.R
\name{load_gbrs}
\alias{load_gbrs}
\title{Load a GBRS Model from JSON File}
\usage{
load_gbrs(filepath)
}
\arguments{
\item{filepath}{Path to the model file}
}
\value{
A gbrs model object
}
\description{
Load a GBRS Model from JSON File
}
