% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_io.R
\name{load_predictions}
\alias{load_predictions}
\title{Load Predictions from JSON File}
\usage{
load_predictions(filepath)
}
\arguments{
\item{filepath}{Path to predictions file}
}
\value{
Numeric vector of predictions
}
\description{
Load Predictions from JSON File
}
