% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ascii.horizontal}
\alias{print.ascii.horizontal}
\title{Print GBRS Model as Horizontal ASCII Table}
\usage{
\method{print}{ascii.horizontal}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"gbrs"} returned by \code{\link{gbrs}}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print GBRS Model as Horizontal ASCII Table
}
