% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.latex.horizontal}
\alias{print.latex.horizontal}
\title{Print GBRS Model as Horizontal LaTeX Table}
\usage{
\method{print}{latex.horizontal}(x, ...)
}
\description{
Print GBRS Model as Horizontal LaTeX Table
}
