% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.md}
\alias{print.md}
\title{Print GBRS Model as Markdown Table}
\usage{
\method{print}{md}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"gbrs"} returned by \code{\link{gbrs}}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the Markdown code as a character string.
}
\description{
S3 method to generate a Markdown table representation of the GBRS model score.
}
\examples{
model <- gbrs(mpg ~ wt + hp, data = mtcars, n_max = 10, n_quantiles = 5)
print(model, "md")

}
