% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.vertical}
\alias{print.vertical}
\title{Print GBRS Model in Legacy Vertical Format}
\usage{
\method{print}{vertical}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"gbrs"} returned by \code{\link{gbrs}}.}

\item{...}{Additional arguments passed to print.}
}
\description{
Prints the GBRS model using the legacy vertical format.
}
