% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_io.R
\name{save_gbrs}
\alias{save_gbrs}
\title{Save a GBRS Model to JSON File}
\usage{
save_gbrs(model, filepath)
}
\arguments{
\item{model}{A gbrs model object}

\item{filepath}{Path to save the model file}
}
\description{
Save a GBRS Model to JSON File
}
