% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_io.R
\name{save_predictions}
\alias{save_predictions}
\title{Save Predictions to JSON File}
\usage{
save_predictions(predictions, filepath)
}
\arguments{
\item{predictions}{Numeric vector of predictions}

\item{filepath}{Path to save predictions}
}
\description{
Save Predictions to JSON File
}
