% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gbrs-package}
\alias{gbrs}
\alias{gbrs-package}
\title{gbrs: Gradient Boosted Risk Scores}
\description{
GBRS (gradient boosted risk scoring) is a library that generates risk scores.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/costa/gbrs}
  \item Report bugs at \url{https://github.com/costa/gbrs/issues}
}

}
\author{
\strong{Maintainer}: Costa Georgantas \email{costa@example.com}

}
\keyword{internal}
