% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ascii.horizontal}
\alias{print.ascii.horizontal}
\title{Print GBRS Model as Horizontal ASCII Table}
\usage{
\method{print}{ascii.horizontal}(x, ...)
}
\description{
Print GBRS Model as Horizontal ASCII Table
}
