% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.latex.horizontal}
\alias{print.latex.horizontal}
\title{Print GBRS Model as Horizontal LaTeX Table}
\usage{
\method{print}{latex.horizontal}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"gbrs"} returned by \code{\link{gbrs}}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print GBRS Model as Horizontal LaTeX Table
}
