% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.md.horizontal}
\alias{print.md.horizontal}
\title{Print GBRS Model as Horizontal Markdown Table}
\usage{
\method{print}{md.horizontal}(x, ...)
}
\description{
Print GBRS Model as Horizontal Markdown Table
}
