% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.gbrs_bootstrap}
\alias{summary.gbrs_bootstrap}
\title{Get Bootstrap Weight Statistics}
\usage{
\method{summary}{gbrs_bootstrap}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"gbrs_bootstrap"}.}
}
\value{
A data frame with columns: feature, threshold, mean, std
}
\description{
Extract mean and standard deviation for each weight from bootstrap results.
}
